pyproject.toml
[tool.poetry]
name = "discord-notify"
version = "0.1.0"
description = ""
authors = ["trasta <trasta298@gmail.com>"]
readme = "README.md"
packages = [{include = "discord_notify"}]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^2.2.2"
requests = "^2.28.2"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
